Corrección en tipo de valor de Id para Comodato y Remito debido al limite de int32 respecto a long; idem en BD.

+++ BE/Remito.cs
@@ -11,7 +11,7 @@ namespace BE
 
         public DateTime Fecha { get; set; }
 
-        public int Numero { get; set; }
+        public long Numero { get; set; }
